chore: refactor common functions into common file
