+++ talentmap_api/fsbid/services/common.py
@@ -1,4 +1,5 @@
 import requests
+import re
 import logging
 
 from urllib.parse import urlencode
@@ -26,3 +27,76 @@ def get_pagination(query, count, base_url, host=None):
 def convert_multi_value(val):
     if val is not None:
         return val.split(',')
+
+
+# Pattern for extracting language parts from a string. Ex. "Spanish (3/3)"
+LANG_PATTERN = re.compile("(.*?)\(.*\)\s(\d)/(\d)")
+
+def parseLanguage(lang):
+    '''
+    Parses a language string from FSBid and turns it into what we want
+    The lang param comes in as something like "Spanish (3/3)"
+    '''
+    if lang:
+        match = LANG_PATTERN.search(lang)
+        if match:
+            language = {}
+            language["language"] = match.group(1)
+            language["reading_proficiency"] = match.group(2)
+            language["spoken_proficiency"] = match.group(3)
+            language["representation"] = match.group(0).rstrip()
+            return language
+
+
+def post_values(query):
+    '''
+    Handles mapping locations and groups of locations to FSBid expected params
+    '''
+    results = []
+    if query.get("position__post__in"):
+        post_ids = query.get("position__post__in").split(",")
+        location_codes = Post.objects.filter(id__in=post_ids).values_list("_location_code", flat=True)
+        results = results + list(location_codes)
+    if len(results) > 0:
+        return results
+
+
+def bureau_values(query):
+    '''
+    Gets the ids for the functional/regional bureaus and maps to codes and their children
+    '''
+    results = []
+    # functional bureau filter
+    if query.get("org_has_groups"):
+        func_bureaus = query.get("org_has_groups").split(",")
+        func_org_codes = OrganizationGroup.objects.filter(id__in=func_bureaus).values_list("_org_codes", flat=True)
+        # Flatten _org_codes
+        func_bureau_codes = [item for sublist in func_org_codes for item in sublist]
+        results = results + list(func_bureau_codes)
+    # Regional bureau filter
+    if query.get("position__bureau__code__in"):
+        regional_bureaus = query.get("position__bureau__code__in").split(",")
+        reg_org_codes = Organization.objects.filter(Q(code__in=regional_bureaus) | Q(_parent_organization_code__in=regional_bureaus)).values_list("code", flat=True)
+        results = results + list(reg_org_codes)
+    if len(results) > 0:
+        return results
+
+def overseas_values(query):
+    '''
+    Maps the overseas/domestic filter to the proper value
+    '''
+    if query.get("is_domestic") == "true":
+        return "D"
+    if query.get("is_domestic") == "false":
+        return "O"
+sort_dict = {
+    "position__title": "pos_title_desc",
+    "position__grade": "pos_grade_code",
+    "position__bureau": "bureau_desc",
+    "ted": "ted",
+    "position__position_number": "position"
+}
+
+def sorting_values(sort):
+    if sort is not None:
+        return sort_dict.get(sort, None)
